Created a website with bootstrap
It have an API to display random images from an choosen tag and an API to display and calculate how much an BRL worth in BTC (or better, in satoshis) alongside with an JS script to do this math work. :D
I'm not gonna lie, loved to saw it working!